Add tooltip to demonstration page lines to display individual line parse results
This uses CSS anchor positioning, which is still pretty new and not supported in all browsers.
Browsers that don't support it yet do not display the tooltip.
Since this is just a demonstration page and CSS anchor positioning will soon be supported in all
up-to-date browsers, I didn't bother installing a polyfill.